requirements.txt
amqp==2.4.2
billiard==3.5.0.5
celery==4.2.2
certifi==2019.3.9
kombu==4.3.0
psycopg2-binary==2.7.7
pytz==2018.9
PyYAML==5.1
redis==3.2.1
vine==1.3.0
